GARBAGE COLLECTION YALLLL WOOOOOOOO
Install libgc and go to sinscm and set your path (youll find the comment) and itll work i dont promise but i want to.

NOTES:

* sinscm interface takes more arguments now, an input-type (-j) and an output-type (-t), RTFM.

* CHANGES TO SinObj

SinObj now has a third member, a pointer. So an Object can either be a pointer,
or an atomic value, and it needs be typed that way.

So an Int will be {type:Int, value:THEVAL, ptrvalue:NULL}

And a Vector will be {type:Vector, value:0, ptrvalue:VECTORPOINTER}

This way the GC is able to mark the vectors internal pointers as used, becuase
we dont coerce from pointer to atomic anymore, only pointer to pointer (u64*).

* Fix a lot of -Wall -Weverything warnings, damnit!

* Implement eq?(mostly ?).

* unwrap_* now takes a string argument for error message formatting. Its useful!

* Made emitted LLVM code much prettier by fixing freevariable whitespace excess.