alloc_atomic support! Now we have a cleaner object model with more efficient allocations
We now use GC_MALLOC_ATOMIC to allocate objects that are atomic, like Int or Bool.
This is more efficent, as the GC will not try to follow pointers in those objects.

This is necessary because we changed the object layout to be {Type,Ptr} where Ptr is a u64*.
Of course, Ints are not pointers, but for compactness in the object, we type-cast it.
Using GC_MALLOC_ATOMIC means we can safely cast it to a pointer without the GC trying to reach into it and get garbage memory.

Also some cleanups.